moved sandbox metrics endpoint to separate file
